Fix de las colisiones

+++ AnimalCooking/PlayersAdder.cpp
@@ -49,8 +49,8 @@ void PlayersAdder::players_initializeTransform(size_t player, const double casil
 	t->setPos(Vector2D(jsonLevel["Players"][player]["pos"]["x"].as_double() * casillaLength,
 		jsonLevel["Players"][player]["pos"]["y"].as_double() * casillaLength));
 
-	t->setHitboxOffset(t->getW() * 0.1, t->getH() / 2);
-	t->setHitboxSize(t->getW() * 0.8, t->getH() / 2);
+	t->setHitboxOffset(int(t->getW() * 0.1), int(t->getH() / 2));
+	t->setHitboxSize(int(t->getW() * 0.8), int(t->getH() / 2));
 }
 
 constexpr unsigned int str2int(const char* str, int h = 0)
